+++ src/StripIdFixModule.hpp
@@ -1,13 +1,13 @@
 #pragma once
 
 struct StripIdFixModule {
-	const std::map<int, ModuleWidget*>* idFixMap = NULL;
+	const std::map<int64_t, ModuleWidget*>* idFixMap = NULL;
 
-	void idFixDataFromJson(const std::map<int, ModuleWidget*>& moduleIdMapping) {
+	void idFixDataFromJson(const std::map<int64_t, ModuleWidget*>& moduleIdMapping) {
 		this->idFixMap = &moduleIdMapping;
 	}
 
-	int idFix(int moduleId) {
+	int64_t idFix(int64_t moduleId) {
 		if (!this->idFixMap) return moduleId;
 		auto it = this->idFixMap->find(moduleId);
 		if (it == this->idFixMap->end()) return -1;
